refactor: extract ref setter
