+++ src/InfiniteScroll/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './getValidScrollValue';
 export * from './getMultiplierSign';
+export * from './getScrollingContainerRef';
+export * from './isValidScrollingContainerRef';
