+++ src/__tests__/InfiniteScroll.test.ts
@@ -264,7 +264,7 @@ describe('InfiniteScroll', () => {
         expect(container?.scroll).toEqual(undefined);
       });
 
-      it('should NOT call "onScroll" callback', async () => {
+      it('should call "onScroll" callback', async () => {
         const instanceProps: InfiniteScrollProps = {
           ...mockInfiniteScrollProps,
           onScroll: () => {},
@@ -280,7 +280,7 @@ describe('InfiniteScroll', () => {
 
         await update();
 
-        expect(spyOnScroll).not.toHaveBeenCalled();
+        expect(spyOnScroll).toHaveBeenCalled();
       });
     });
   });
